# Conflicts:
#	Android/app/src/debug/java/com/didichuxing/doraemondemo/MainDebugActivity.java
#	Android/app/src/main/java/com/didichuxing/doraemondemo/util/FrescoUtil.java

+++ Android/doraemonkit/src/main/java/com/didichuxing/doraemonkit/config/PerformanceSpInfoConfig.java
@@ -105,7 +105,7 @@ public class PerformanceSpInfoConfig {
      * @param context
      * @param threshold
      */
-    public static float getLargeImgMemoryThreshold(Context context, float threshold) {
+    public static double getLargeImgMemoryThreshold(Context context, float threshold) {
         return SharedPrefsUtil.getFloat(context, SharedPrefsKey.LARGE_IMG_MEMORY_THRESHOLD, threshold);
     }
 
@@ -126,7 +126,7 @@ public class PerformanceSpInfoConfig {
      * @param context
      * @param threshold
      */
-    public static float getLargeImgFileThreshold(Context context, float threshold) {
+    public static double getLargeImgFileThreshold(Context context, float threshold) {
         return SharedPrefsUtil.getFloat(context, SharedPrefsKey.LARGE_IMG_FILE_THRESHOLD, threshold);
     }
